requirements.txt
PyMySQL>=1.0.2
SQLAlchemy>=1.4.39
python-dotenv>=0.20.0
discord>=2.2.2
pytz>=2022.1
python-dateutil>=2.8.2
gspread>=5.7.2
oauth2client>=4.1.3
timer>=0.2.2
Levenshtein>=0.21.0
asyncio>=3.4.3